requirements.txt
loguru~=0.5.3
pyTelegramBotAPI~=4.3.1
python-dotenv~=0.19.2
requests~=2.26.0